:recycle: refactor : 애플리케이션 진입점 및 라우팅 구조 구성
