+++ index.html
@@ -9,6 +9,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.jsx"></script>
+    <script src="/src/app/main.jsx" type="module"></script>
   </body>
 </html>
