+++ src/app/main.jsx
@@ -1,5 +1,5 @@
 import {createRoot} from 'react-dom/client'
-import './index.css'
+import '../index.css'
 import {BrowserRouter} from 'react-router-dom';
 import App from './App.jsx'
 import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
